+++ R/itemstats.R
@@ -25,6 +25,7 @@
 #' Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
 #' \doi{10.18637/jss.v048.i06}
 #' @keywords data
+#' @seealso \code{\link{empirical_plot}}
 #' @export
 #' @examples
 #'