+++ R/mirt.R
@@ -631,7 +631,7 @@
 #'   \code{\link{expand.table}}, \code{\link{key2binary}}, \code{\link{mod2values}},
 #'   \code{\link{extract.item}}, \code{\link{iteminfo}}, \code{\link{testinfo}},
 #'   \code{\link{probtrace}}, \code{\link{simdata}}, \code{\link{averageMI}},
-#'   \code{\link{fixef}}, \code{\link{extract.mirt}}
+#'   \code{\link{fixef}}, \code{\link{extract.mirt}}, \code{\link{itemstats}}
 #'
 #' @references
 #'