+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - Added `itemstats()` function to give basic item information statistics
 
-- Item-EFA models now autmatically flip negative signs in rotate solutions
+- Item-EFA models now automatically flips negative signs in rotate solutions
   (e.g., via `summary()`) according to the sign of the largest observed 
   loading (allows easier interpretation of the resulting correlation matrix)